move details utils and const
