+++ corehq/apps/app_manager/suite_xml/const.py
@@ -1,0 +1,6 @@
+FIELD_TYPE_ATTACHMENT = 'attachment'
+FIELD_TYPE_INDICATOR = 'indicator'
+FIELD_TYPE_LOCATION = 'location'
+FIELD_TYPE_PROPERTY = 'property'
+FIELD_TYPE_LEDGER = 'ledger'
+FIELD_TYPE_SCHEDULE = 'schedule'
